test: refactor test collections for do-complete
