+++ packages/language-services/src/features/__tests__/do-complete-modules.test.ts
@@ -84,56 +84,6 @@ test("should suggest symbol from a different document via @use", async () => {
 	);
 });
 
-test("should suggest symbol from a different document via @use when in string interpolation", async () => {
-	ls.configure({
-		completionSettings: {
-			suggestFromUseOnly: true,
-		},
-	});
-
-	const one = fileSystemProvider.createDocument("$primary: limegreen;", {
-		uri: "one.scss",
-	});
-	const two = fileSystemProvider.createDocument(
-		['@use "./one";', '.a { background: url("/#{one.'],
-		{
-			uri: "two.scss",
-		},
-	);
-
-	// emulate scanner of language service which adds workspace documents to the cache
-	ls.parseStylesheet(one);
-	ls.parseStylesheet(two);
-
-	const { items } = await ls.doComplete(two, Position.create(1, 29));
-	assert.ok(items.find((annotation) => annotation.label === "$primary"));
-});
-
-test("should suggest symbol from a different document via @use when in @return", async () => {
-	ls.configure({
-		completionSettings: {
-			suggestFromUseOnly: true,
-		},
-	});
-
-	const one = fileSystemProvider.createDocument("$primary: limegreen;", {
-		uri: "one.scss",
-	});
-	const two = fileSystemProvider.createDocument(
-		['@use "./one";', "@function test() { @return one."],
-		{
-			uri: "two.scss",
-		},
-	);
-
-	// emulate scanner of language service which adds workspace documents to the cache
-	ls.parseStylesheet(one);
-	ls.parseStylesheet(two);
-
-	const { items } = await ls.doComplete(two, Position.create(1, 31));
-	assert.ok(items.find((annotation) => annotation.label === "$primary"));
-});
-
 test("should not suggest symbols from a module used by the one we use", async () => {
 	ls.configure({
 		completionSettings: {
